// src/services/model/ApiBook.ts
/**
 * unknown
 * unknown
 *
 * OpenAPI spec version: last
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

/**
  * @property `[errno]` 
  * @property `[errmsg]` 
  * @property `[count]` 
  * @property `[data]` 
  */
export interface ApiBook {
  "errno"?: number;
  "errmsg"?: string;
  "count"?: number;
  "data"?: Array<models.ApiBookData>;
}